+++ src/r_data.c
@@ -534,8 +534,8 @@ static void R_InitSpriteLumps(void)
     {
         mobjinfo[MT_HEAD].flags2 |= MF2_DONTMAP;
         mobjinfo[MT_INV].flags2 &= ~MF2_TRANSLUCENT_33;
-        mobjinfo[MT_INS].flags2 &= ~(MF2_TRANSLUCENT_33 | MF2_FLOATBOB | MF2_NOFOOTCLIP);
-        mobjinfo[MT_MISC14].flags2 &= ~(MF2_FLOATBOB | MF2_NOFOOTCLIP);
+        mobjinfo[MT_INS].flags2 &= ~(MF2_TRANSLUCENT_33 | MF2_FLOATBOB);
+        mobjinfo[MT_MISC14].flags2 &= ~MF2_FLOATBOB;
         mobjinfo[MT_BFG].flags2 &= ~MF2_TRANSLUCENT;
         mobjinfo[MT_HEAD].blood = MT_BLOOD;
         mobjinfo[MT_BRUISER].blood = MT_BLOOD;