+++ releasenotes.md
@@ -13,6 +13,7 @@
 * The [*DOOM II* monster exclusion bug](https://doomwiki.org/wiki/Doom_II_monster_exclusion_bug) present in *Vanilla DOOM* has been fixed.
 * The `vid_widescreen` CVAR can now be changed in the console when not during a game.
 * Intermission and finale text is now also output to the console.
+* The bottom of lost souls are now clipped when touching a liquid sector.
 
 ---